+++ server/.env
@@ -1,3 +1,3 @@
 TEAM_MEMBERS="Aecio, Bonnie, Dhruv, Hakeem, Thiago"
-AWS_SECRET_ACCESS_KEY = ""
-AWS_ACCESS_KEY_ID = ""
+AWS_SECRET_ACCESS_KEY = "9YAtUuiMi31UI5XTvJx298w62u0qtBUNHybA72Cu"
+AWS_ACCESS_KEY_ID = "AKIAY24KUFXNQW3YNBNI"
