+++ server/models/polls.js
@@ -2,37 +2,42 @@ const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
 const PollSchema = new Schema({
-	title: {
-		type: String,
-		required: true,
-	},
-	userId: {
-		type: Number,
-		required: true,
-	},
-	date: {
-		type: Date,
-		default: Date.now,
-	},
-	images: [
-		{
-			url: {
-				type: String,
-				required: true,
-			},
-		},
-	],
+  title: {
+    type: String,
+    required: true,
+  },
+  userId: {
+    type: Number,
+    required: true,
+  },
+  date: {
+    type: Date,
+    default: Date.now,
+  },
+  images: [
+    {
+      url: {
+        type: String,
+        required: true,
+      },
+    },
+  ],
 });
 
 PollSchema.methods.addImage = function (imageLocation) {
-	this.images.push({ url: imageLocation });
+  this.images.push({ url: imageLocation });
 };
 
 PollSchema.methods.updateImage = function (imageLocation, id) {
-	const imageObj = this.images.find((image) => String(image._id) === id);
-	imageObj["url"] = imageLocation;
-	// Saving the parent, because the child cannot be saved
-	this.save();
+  const imageObj = this.images.find((image) => String(image._id) === id);
+  imageObj["url"] = imageLocation;
+  // Saving the parent, because the child cannot be saved
+  this.save();
+};
+
+PollSchema.methods.updateTitle = function (title) {
+  this.title = title;
+  this.save();
 };
 
 module.exports = Poll = mongoose.model("polls", PollSchema);
